bump sbt-coursier to latest stable version
this makes classutil buildable on JDK 11.  this is useful in the
context of the Scala community build, to help us test Scala
on JDK 11.

I also threw in an sbt version bump, since it's the sbt version
the community build is using.

both changes are also just generally desirable afaik.